mofan-realtime-py-h5/backend/requirements.txt
fastapi
uvicorn
websockets
python-dotenv
aiohttp
